Make GPX importer more robust.
Handle the case where the first track point does not have a time
without segfaulting. (issue #316, #192)

Don't throw an exception on a negative time change. It isn't the end
of the world, just put a warning in the log and move on (without
updating the speed, since we really don't have a clue now). (issue #186)

Reset the element content even when we aren't in a track element.  GPX
can have elements outside of the <trk>.  This was causing problems
such as content in a <metadata> element before the <trk> being
included in the track name.

There could still be weirdness in the stats if some of the points have
a time element while others don't.  Hard to say what the best course
of action is in that case though, and may involve changes to the
TripStatisticsBuilder.